PO-5935 : NLM-CTRP Integration: Daily retrieval from CT.GOV of trial status and other data for Industrial Trials :: CR : COPPA-793
Added nightly job to import update industrial and consortia trials from CT.Gov and update them in CTRP : 

(1) CTGovSyncNightlyServiceLocal/CTGovSyncNightlyServiceBeanLocal : 
(a) Queries the industrial and consortia trials in CTRP with NCT identifier. 
(b) For each of the trials checks if there is a corresponding entry in CTGovImportLog table. 
 If so it compares the last changed date in CT.Gov with created date of the log entry.If the last 
changed date in CT.Gov is more recent then it imports the trial from CT.Gov and updates the same in 
CTRP. Else skips the update. 
(ii) If there is no corresponding entry in CTGovImportLog table, it imports the trial from CT.Gov and 
updates the same in CTRP. Else skips the update. 

(2) quartz_jobs.xml - Added CTGovSyncNightlyServiceLocal as the nightly job. 

(3) Updated the ServiceLocator classes.

(4) CTGovSyncNightlyServiceLocalTest - Exercises CTGovSyncNightlyServiceBeanLocal.

git-svn-id: https://ncisvn.nci.nih.gov/svn/coppa/trunk/code@15607 553898ca-32d5-487a-af8b-6cf4ca579ab4